reorganization of the project with different scripts for different steps
