+++ preprocessing.py
@@ -1,0 +1,17 @@
+## This module defines the functions used for preprocessing the data in order. It should include at least the following functions
+
+def determinePrefix():
+    """ Checks for regular patterns that can be considered prefixes
+    """
+    pass
+
+
+def determineSuffix():
+    """ Checks for regular patterns that can be considered suffixes
+    """
+    pass
+
+def checkTranliterationDifferences():
+    """ Checks for different ways to transliterate similar sounds
+    """
+    pass
